Integrate PR feedback: inline test fixture, use a settings constant for min # of revisions.
